# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  reseller: "reseller",
  client: "client",
} as const;

export type UserService = (typeof UserService)[keyof typeof UserService] | null;

export const UserService = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  service?: UserService;
  serverId?: number | null;
  packageId?: number | null;
  active: boolean;
  expiresAt?: string | null;
  resellerId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  reseller: "reseller",
  client: "client",
} as const;

export type CreateUserBodyService =
  (typeof CreateUserBodyService)[keyof typeof CreateUserBodyService];

export const CreateUserBodyService = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
  service?: CreateUserBodyService;
  serverId?: number;
  packageId?: number;
  expiresAt?: string;
  resellerId?: number;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  reseller: "reseller",
  client: "client",
} as const;

export type UpdateUserBodyService =
  (typeof UpdateUserBodyService)[keyof typeof UpdateUserBodyService];

export const UpdateUserBodyService = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  password?: string;
  role?: UpdateUserBodyRole;
  service?: UpdateUserBodyService;
  serverId?: number;
  packageId?: number;
  active?: boolean;
  expiresAt?: string;
  resellerId?: number;
}

export interface ResellerWithStats {
  id: number;
  name: string;
  email: string;
  active: boolean;
  createdAt: string;
  clientCount: number;
}

export interface CreateResellerBody {
  name: string;
  email: string;
  password: string;
}

export interface UpdateResellerBody {
  name?: string;
  email?: string;
  password?: string;
  active?: boolean;
}

export type ServerType = (typeof ServerType)[keyof typeof ServerType];

export const ServerType = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface Server {
  id: number;
  name: string;
  url: string;
  type: ServerType;
  apiKey: string;
  active: boolean;
  userCount: number;
  createdAt: string;
}

export type CreateServerBodyType =
  (typeof CreateServerBodyType)[keyof typeof CreateServerBodyType];

export const CreateServerBodyType = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface CreateServerBody {
  name: string;
  url: string;
  type: CreateServerBodyType;
  apiKey: string;
  active?: boolean;
}

export type UpdateServerBodyType =
  (typeof UpdateServerBodyType)[keyof typeof UpdateServerBodyType];

export const UpdateServerBodyType = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface UpdateServerBody {
  name?: string;
  url?: string;
  type?: UpdateServerBodyType;
  apiKey?: string;
  active?: boolean;
}

export type ServerTestResultChecksItemStatus =
  (typeof ServerTestResultChecksItemStatus)[keyof typeof ServerTestResultChecksItemStatus];

export const ServerTestResultChecksItemStatus = {
  ok: "ok",
  fail: "fail",
  skip: "skip",
} as const;

export type ServerTestResultChecksItem = {
  name: string;
  status: ServerTestResultChecksItemStatus;
  detail?: string | null;
};

export interface ServerTestResult {
  success: boolean;
  serverName?: string | null;
  version?: string | null;
  responseMs?: number | null;
  checks: ServerTestResultChecksItem[];
  error?: string | null;
}

export type PackageServicesItem =
  (typeof PackageServicesItem)[keyof typeof PackageServicesItem];

export const PackageServicesItem = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface Package {
  id: number;
  name: string;
  price: number;
  durationDays: number;
  services: PackageServicesItem[];
  description?: string | null;
  active: boolean;
  userCount: number;
  createdAt: string;
}

export type CreatePackageBodyServicesItem =
  (typeof CreatePackageBodyServicesItem)[keyof typeof CreatePackageBodyServicesItem];

export const CreatePackageBodyServicesItem = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface CreatePackageBody {
  name: string;
  price: number;
  durationDays: number;
  services: CreatePackageBodyServicesItem[];
  description?: string;
  active?: boolean;
}

export type UpdatePackageBodyServicesItem =
  (typeof UpdatePackageBodyServicesItem)[keyof typeof UpdatePackageBodyServicesItem];

export const UpdatePackageBodyServicesItem = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export interface UpdatePackageBody {
  name?: string;
  price?: number;
  durationDays?: number;
  services?: UpdatePackageBodyServicesItem[];
  description?: string;
  active?: boolean;
}

export type DashboardStatsClientsByService = {
  plex: number;
  jellyfin: number;
  emby: number;
};

export interface DashboardStats {
  totalClients: number;
  totalResellers: number;
  totalServers: number;
  activeServers: number;
  totalPackages: number;
  activeDemos: number;
  clientsByService: DashboardStatsClientsByService;
  recentUsers: User[];
}

export interface StreamingSession {
  user: string;
  title: string;
  progress: number;
  quality: string;
}

export interface StreamingSessions {
  plex: number;
  jellyfin: number;
  emby: number;
  plexSessions: StreamingSession[];
  jellyfinSessions: StreamingSession[];
  embySessions: StreamingSession[];
}

export type ListUsersParams = {
  role?: ListUsersRole;
  service?: ListUsersService;
  search?: string;
  resellerId?: number;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  reseller: "reseller",
  client: "client",
} as const;

export type ListUsersService =
  (typeof ListUsersService)[keyof typeof ListUsersService];

export const ListUsersService = {
  plex: "plex",
  jellyfin: "jellyfin",
  emby: "emby",
} as const;

export type ListResellersParams = {
  search?: string;
};