# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamSync Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a package by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageBody"
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /resellers:
    get:
      operationId: listResellers
      tags: [resellers]
      summary: List all resellers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of resellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResellerWithStats"
    post:
      operationId: createReseller
      tags: [resellers]
      summary: Create a new reseller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResellerBody"
      responses:
        "201":
          description: Reseller created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /resellers/{id}:
    put:
      operationId: updateReseller
      tags: [resellers]
      summary: Update a reseller
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResellerBody"
      responses:
        "200":
          description: Reseller updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteReseller
      tags: [resellers]
      summary: Delete a reseller
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /resellers/{id}/toggle:
    post:
      operationId: toggleReseller
      tags: [resellers]
      summary: Toggle reseller active state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reseller toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, reseller, client]
        service:
          type: string
          enum: [plex, jellyfin, emby]
          nullable: true
        serverId:
          type: integer
          nullable: true
        packageId:
          type: integer
          nullable: true
        active:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        resellerId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, active, createdAt]
...
    ResellerWithStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
        clientCount:
          type: integer
      required: [id, name, email, active, createdAt, clientCount]
    CreateResellerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]
    UpdateResellerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        active:
          type: boolean
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        services:
          type: array
          items:
            type: string
            enum: [plex, jellyfin, emby]
        description:
          type: string
          nullable: true
        active:
          type: boolean
        userCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, price, durationDays, services, active, userCount, createdAt]
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        services:
          type: array
          items:
            type: string
            enum: [plex, jellyfin, emby]
        description:
          type: string
        active:
          type: boolean
      required: [name, price, durationDays, services]
    UpdatePackageBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        services:
          type: array
          items:
            type: string
            enum: [plex, jellyfin, emby]
        description:
          type: string
        active:
          type: boolean
...