# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserService } from "./userService";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  service?: UserService;
  serverId?: number | null;
  packageId?: number | null;
  active: boolean;
  expiresAt?: string | null;
  resellerId?: number | null;
  createdAt: string;
}